% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camsizerxle.r
\name{camsizerxle}
\alias{camsizerxle}
\title{Read in data from CAMSIZER machine output}
\usage{
camsizerxle(path, outpath, outfile)
}
\arguments{
\item{path}{is the path where the CAMSIZER data is stored.}

\item{outpath}{is where you want the new data to go.}

\item{outfile}{is the name of the output file.}
}
\value{
saves and returns a set of files with the desired paths.
}
\description{
\code{camsizerxle} provides a tool to read in xle data as output by te CAMSIZER and export it as a more common csv or rda file format.
}
\examples{
camsizerxle(path,outpath,outfile)
}
